auth/perm refactor

+++ app/controllers/invitation.ts
@@ -10,7 +10,7 @@ import type {
 
 export async function create(invitation: NewInvitation): Promise<Invitation> {
   return await db
-    .insertInto('invitation')
+    .insertInto('auth.invitation')
     .values({ ...invitation })
     .returningAll()
     .executeTakeFirstOrThrow();
@@ -20,7 +20,7 @@ export async function getById(
   id: InvitationId,
 ): Promise<Invitation | undefined> {
   return await db
-    .selectFrom('invitation')
+    .selectFrom('auth.invitation')
     .where('id', '=', id)
     .where('expiresAt', '>', new Date())
     .selectAll()
@@ -31,7 +31,7 @@ export async function getByEmail(
   email: string,
 ): Promise<Invitation | undefined> {
   return await db
-    .selectFrom('invitation')
+    .selectFrom('auth.invitation')
     .where('email', '=', email)
     .where('isUsed', '=', false)
     .where('expiresAt', '>', new Date())
@@ -41,7 +41,7 @@ export async function getByEmail(
 
 export async function getAllByOrgId(orgId: OrgId): Promise<Invitation[]> {
   return await db
-    .selectFrom('invitation')
+    .selectFrom('auth.invitation')
     .where('orgId', '=', orgId)
     .where('isUsed', '=', false)
     .selectAll()
@@ -50,12 +50,12 @@ export async function getAllByOrgId(orgId: OrgId): Promise<Invitation[]> {
 
 export async function update(id: InvitationId, updateWith: InvitationUpdate) {
   return await db
-    .updateTable('invitation')
+    .updateTable('auth.invitation')
     .set(updateWith)
     .where('id', '=', id)
     .execute();
 }
 
 export async function deleteInvitation(id: InvitationId) {
-  return await db.deleteFrom('invitation').where('id', '=', id).execute();
+  return await db.deleteFrom('auth.invitation').where('id', '=', id).execute();
 }
